thinking_object_oriented/src/requirements.txt
ruamel.yaml==0.15.35
requests==2.2.1
